// src/app/models/house.model.ts
export class House {
    url: string;
    name: string;
    region: string;
    coatOfArms: string; // ??
    words: string;
    title: Array<string>;
    seat: Array<string>;
    currentLord: string;
    heir: string;
    overlord: string; // ba vuong
    founded: string;
    founder: string;
    diedOut: string;
    ancestralWeapon: Array<string>; // vũ khí
    cadetBranch: Array<string>;
    swornMember: Array<string>;
}